Wizard: Install and activate Jetpack if needed
Add function to install & activate Jetpack as needed.
Props to @jeffstieler for the idea, and the implementation/code:
https://github.com/Automattic/woocommerce-services/pull/1031#discussion_r122505316
https://github.com/Automattic/jetpack/blob/master/3rd-party/woocommerce-services.php#L101

Add a list of errors and a function to get the correct error message.

After clicking the connect Jetpack button, make sure that
Jetpack is installed and activated.

On error, redirect to the activate page with a query arg.